commit onboarding on initial start only
