+++ locales_dev/en/translation.json
@@ -110,10 +110,12 @@
   "Error": "Error",
   "Example": "Example",
   "Expires:": "Expires:",
+  "Get Help": "Get Help",
   "GitHub Users": "GitHub Users",
   "Go Back": "Go Back",
   "Heading": "Heading",
   "Hello! What can Metecho help you do today?": "Hello! What can Metecho help you do today?",
+  "Help Walkthrough": "Help Walkthrough",
   "Home": "Home",
   "I Agree": "I Agree",
   "I want to Help": "I want to Help",
@@ -154,7 +156,9 @@
   "Org:": "Org:",
   "Package Directories": "Package Directories",
   "Pending review": "Pending review",
+  "Plan Walkthrough": "Plan Walkthrough",
   "Planned": "Planned",
+  "Play Walkthrough": "Play Walkthrough",
   "Please close this message and try your action again.": "Please close this message and try your action again.",
   "Post-Install Directories": "Post-Install Directories",
   "Pre-Install Directories": "Pre-Install Directories",
