+++ src/js/components/header.tsx
@@ -1,14 +1,15 @@
 import Dropdown from '@salesforce/design-system-react/components/menu-dropdown';
 import PageHeader from '@salesforce/design-system-react/components/page-header';
 import PageHeaderControl from '@salesforce/design-system-react/components/page-header/control';
+import i18n from 'i18next';
 import React, { useCallback, useState } from 'react';
 import { CallBackProps, STATUS } from 'react-joyride';
 import { useSelector } from 'react-redux';
 import { Link } from 'react-router-dom';
 
-import backpackSvg from '!raw-loader!~img/backpack.svg';
-import mapSvg from '!raw-loader!~img/map.svg';
-import seesawSvg from '!raw-loader!~img/seesaw.svg';
+import backpackSvg from '!raw-loader!~img/backpack-sm.svg';
+import mapSvg from '!raw-loader!~img/map-sm.svg';
+import seesawSvg from '!raw-loader!~img/seesaw-sm.svg';
 import Errors from '~js/components/apiErrors';
 import OfflineAlert from '~js/components/offlineAlert';
 import Toasts from '~js/components/toasts';
@@ -18,7 +19,6 @@ import UserInfo from '~js/components/user/info';
 import { selectSocketState } from '~js/store/socket/selectors';
 import { selectUserState } from '~js/store/user/selectors';
 import routes from '~js/utils/routes';
-import i18n from 'i18next';
 
 const Header = () => {
   const user = useSelector(selectUserState);
@@ -37,7 +37,6 @@ const Header = () => {
     <PageHeaderControl className="slds-grid slds-grid_vertical-align-center">
       <div className="slds-col_padded">
         <Dropdown
-          // todo translate
           assistiveText={{ icon: `${i18n.t('Get Help')}` }}
           buttonVariant="icon"
           iconName="question"
@@ -50,22 +49,25 @@ const Header = () => {
             {
               label: `${i18n.t('Play Walkthrough')}`,
               value: 'play',
-              leftIcon: 'kanban',
+              leftIcon: {
+                name: seesawSvg,
+                category: 'utility',
+              },
             },
             {
               label: `${i18n.t('Help Walkthrough')}`,
               value: 'help',
               leftIcon: {
+                name: backpackSvg,
                 category: 'utility',
-                name: 'kanban',
               },
             },
             {
               label: `${i18n.t('Plan Walkthrough')}`,
               value: 'plan',
               leftIcon: {
+                name: mapSvg,
                 category: 'utility',
-                name: 'side_list',
               },
             },
           ]}
