+++ src/js/components/tour/plan.tsx
@@ -1,12 +1,7 @@
 import i18n from 'i18next';
-import React, { useCallback } from 'react';
+import React from 'react';
 import { Trans } from 'react-i18next';
 import Joyride, { CallBackProps, Step } from 'react-joyride';
-import { useDispatch } from 'react-redux';
-
-import { useIsMounted } from '~js/components/utils';
-import { ThunkDispatch } from '~js/store';
-import { onboard } from '~js/store/user/actions';
 
 interface Props {
   handleCallback: (data: CallBackProps) => void;
@@ -28,48 +23,31 @@ const steps: Step[] = [
   },
 ];
 
-const PlanTour = ({ run, handleCallback }: Props) => {
-  const isMounted = useIsMounted();
-  const dispatch = useDispatch<ThunkDispatch>();
-
-  const doOnboard = useCallback(
-    (data: CallBackProps) => {
-      dispatch(onboard()).finally(() => {
-        /* istanbul ignore else */
-        if (isMounted.current) {
-          handleCallback(data);
-        }
-      });
-    },
-    [dispatch, isMounted, handleCallback],
-  );
-
-  return (
-    <Joyride
-      steps={steps}
-      run={run}
-      locale={{
-        back: i18n.t('Back'),
-        close: i18n.t('Close'),
-        last: i18n.t('Close'),
-        next: i18n.t('Next'),
-      }}
-      continuous
-      disableScrolling
-      callback={doOnboard}
-      styles={{
-        options: {
-          textColor: '#080707',
-          primaryColor: '#0070d2',
-        },
-        tooltipContainer: {
-          textAlign: 'left',
-        },
-      }}
-      floaterProps={{
-        disableAnimation: true,
-      }}
-    />
-  );
-};
+const PlanTour = ({ run, handleCallback }: Props) => (
+  <Joyride
+    steps={steps}
+    run={run}
+    locale={{
+      back: i18n.t('Back'),
+      close: i18n.t('Close'),
+      last: i18n.t('Close'),
+      next: i18n.t('Next'),
+    }}
+    continuous
+    disableScrolling
+    callback={handleCallback}
+    styles={{
+      options: {
+        textColor: '#080707',
+        primaryColor: '#0070d2',
+      },
+      tooltipContainer: {
+        textAlign: 'left',
+      },
+    }}
+    floaterProps={{
+      disableAnimation: true,
+    }}
+  />
+);
 export default PlanTour;
